add right system for repostiroy, containers, images, instances, apps

+++ classes/Container.ts
@@ -1,21 +1,29 @@
-import { v4 as uuidv4 } from 'uuid';
-import { promises as fs } from 'fs';
-import fsSync from 'fs';import docker from '../coreDocker';
-import { ContainerDbTable, ContainerDeviceEnvOptionTable, ContainerEnvOptionTable, ContainerResourceEnvOptionTable, ContainerTable, ContainerVariableTable, ContainerVolumeTable, DomainTable } from '../modules/models/tables';
-import EventsObserver from '../modules/events/eventsObserver';
-import ContainerProcess from './ContainerProcess';
-import MegapolosNode from './Node';
-import Image from './Image';
-import Instance, { InstanceRuntimeVariables } from './Instance';
-import { ContainerInput, ContainerResult } from '../types';
-import Volume from './Volume';
+import fsSync from 'fs';
 import { promisify } from 'util';
-import BaseRepository from './BaseRepository';
+import { v4 as uuidv4 } from 'uuid';
+import docker from '../coreDocker';
 import { knex } from '../corePostgres';
+import EventsObserver from '../modules/events/eventsObserver';
+import {
+  ContainerDbTable,
+  ContainerEnvOptionTable,
+  ContainerTable,
+  ContainerVariableTable,
+  ContainerVolumeTable,
+  DomainTable,
+} from '../modules/models/tables';
+import { ContainerResult } from '../types';
+import BaseRepository from './BaseRepository';
 import ContainerDb from './ContainerDb';
-import Domain from './Domain';
+import ContainerProcess from './ContainerProcess';
 import Db from './Db';
 import DbUser from './DbUser';
+import Domain from './Domain';
+import Image from './Image';
+import Instance, { InstanceRuntimeVariables } from './Instance';
+import MegapolosNode from './Node';
+import Volume from './Volume';
+import { resources } from '../src/features/rights/resources_list';
 
 const exec = promisify(require('child_process').exec);
 
@@ -25,18 +33,22 @@ export enum ContainerLifeStatus {
 }
 
 export interface ContainerRuntimeVariables {
-  variables: { [key: string]: string },
-  dbs: { [key: string]: {
-    db: string,
-    host: string,
-    user: string,
-    password: string,
-  } },
-  domain?: string,
-  volumes: { [key: string]: {
-    innerPath: string,
-  } },
-  instance?: InstanceRuntimeVariables,
+  variables: { [key: string]: string };
+  dbs: {
+    [key: string]: {
+      db: string;
+      host: string;
+      user: string;
+      password: string;
+    };
+  };
+  domain?: string;
+  volumes: {
+    [key: string]: {
+      innerPath: string;
+    };
+  };
+  instance?: InstanceRuntimeVariables;
 }
 
 class Container extends BaseRepository<ContainerTable> {
@@ -45,6 +57,7 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async create(data: Partial<ContainerTable>): Promise<ContainerTable> {
+    await this.checkActionAccess(resources.container.actions.create);
     const node = new MegapolosNode(data.node_id);
     let outerPort = await node.getPort();
     if (data.outer_port) {
@@ -55,9 +68,6 @@ class Container extends BaseRepository<ContainerTable> {
     return super.create(data);
   }
 
-  async updateLifeStatus(status: string): Promise<void> {
-    await this.edit({ life_status: status });
-  }
 
   async getDockerContainer() {
     const data = await this.getData();
@@ -66,28 +76,37 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async start() {
+    await this.checkActionAccess(resources.container.actions.manage);
     try {
       // await (await this.getDockerContainer()).start();
     } catch (e) {
       console.trace(e);
-      EventsObserver.listener({ type: 'containerError', data: { containerId: this.id, error: e } });
+      EventsObserver.listener({
+        type: 'containerError',
+        data: { containerId: this.id, error: e },
+      });
     }
-        
-    await this.updateLifeStatus('running');
+
+    await this._updateLifeStatus('running');
   }
 
   async stop() {
+    await this.checkActionAccess(resources.container.actions.manage);
     try {
       // await (await this.getDockerContainer()).stop();
     } catch (e) {
       console.trace(e);
-      EventsObserver.listener({ type: 'containerError', data: { containerId: this.id, error: e } });
+      EventsObserver.listener({
+        type: 'containerError',
+        data: { containerId: this.id, error: e },
+      });
     }
-      
-    await this.updateLifeStatus('stopped');
+
+    await this._updateLifeStatus('stopped');
   }
 
   async edit(data: Partial<ContainerTable>): Promise<ContainerTable> {
+    await this.checkActionAccess(resources.container.actions.edit);
     const entity = await this.getData();
     if (data.outer_port) {
       if (data.outer_port !== entity.outer_port) {
@@ -98,7 +117,8 @@ class Container extends BaseRepository<ContainerTable> {
     return super.edit(data);
   }
 
-  async delete():Promise<boolean> {
+  async delete(): Promise<boolean> {
+    await this.checkActionAccess(resources.container.actions.remove);
     const data = await this.getData();
     if (data.docker_runtime_id) {
       try {
@@ -110,10 +130,14 @@ class Container extends BaseRepository<ContainerTable> {
         // await (await this.getDockerContainer()).remove();
       } catch (e) {
         console.trace(e);
-        EventsObserver.listener({ type: 'containerError', data: { containerId: this.id, error: e } });
+        EventsObserver.listener({
+          type: 'containerError',
+          data: { containerId: this.id, error: e },
+        });
       }
     }
-    const megapolosVolume = MegapolosNode.currentNode.getMegapolosPath() + '/volumes/' + this.id;
+    const megapolosVolume = MegapolosNode.currentNode.getMegapolosPath()
+      + '/volumes/' + this.id;
     if (fsSync.existsSync(megapolosVolume)) {
       MegapolosNode.currentNode.validatePath(megapolosVolume);
       // await fs.rmdir(megapolosVolume, { recursive: true });
@@ -124,11 +148,12 @@ class Container extends BaseRepository<ContainerTable> {
       const volume = volumes[i];
       await this.removeVolume(volume.id);
     }
-  
+
     return super.delete();
   }
 
   async restore() {
+    await this.checkActionAccess(resources.container.actions.manage);
     const container = await this.getData();
     try {
       if (!container.docker_runtime_id) {
@@ -136,41 +161,40 @@ class Container extends BaseRepository<ContainerTable> {
       }
       const containerInfo = await (await this.getDockerContainer()).inspect();
       if (container.life_status === 'running' && !containerInfo.State.Running) {
-        try {
-          await this.start();
-        } catch (e) {
-          console.trace(e);
-          EventsObserver.listener({ type: 'containerError', data: { containerId: this.id, error: e } });
-        }
+        await this.start();
       }
       if (container.life_status === 'stopped' && containerInfo.State.Running) {
-        try {
-          await this.stop();
-        } catch (e) {
-          console.trace(e);
-          EventsObserver.listener({ type: 'containerError', data: { containerId: this.id, error: e } });
-        }
+        await this.stop();
       }
     } catch (e) {
       console.trace(e);
-      EventsObserver.listener({ type: 'containerError', data: { containerId: this.id, error: e } });
+      EventsObserver.listener({
+        type: 'containerError',
+        data: { containerId: this.id, error: e },
+      });
     }
   }
 
   async removeVolume(containerVolumeId: string): Promise<void> {
     const volumes = await new Volume().getVolumesOfContainer(this.id);
-    const volumeContainer = volumes.find((_volume) => _volume.id === containerVolumeId);
+    const volumeContainer = volumes.find((_volume) =>
+      _volume.id === containerVolumeId
+    );
     if (!volumeContainer) {
       throw new Error('Volume not found');
     }
     if (volumeContainer.is_dynamic) {
-      const volumePath = MegapolosNode.currentNode.getMegapolosPath() + '/volumes/' + this.id + '/' + volumeContainer.id;
+      const volumePath = MegapolosNode.currentNode.getMegapolosPath()
+        + '/volumes/' + this.id + '/' + volumeContainer.id;
       MegapolosNode.currentNode.validatePath(volumePath);
       try {
         // await exec(`umount ${volumePath}`);
       } catch (e) {
         console.trace(e);
-        EventsObserver.listener({ type: 'volumeError', data: { containerId: this.id, error: e } });
+        EventsObserver.listener({
+          type: 'volumeError',
+          data: { containerId: this.id, error: e },
+        });
       }
       // await fs.rmdir(volumePath);
     }
@@ -182,10 +206,15 @@ class Container extends BaseRepository<ContainerTable> {
     const volumes = await new Volume().getVolumesOfContainer(container.id);
     container.volumes = volumes;
     const envs = await this.getContainerEnvOptions();
-    container.envs = envs.map((env) => ({ key: env.container_env_name, value: env.container_env_value }));
+    container.envs = envs.map((env) => ({
+      key: env.container_env_name,
+      value: env.container_env_value,
+    }));
     if (container.docker_runtime_id) {
       try {
-        const dockerStatus = (await docker.getContainer(container.docker_runtime_id).inspect()).State.Status;
+        const dockerStatus =
+          (await docker.getContainer(container.docker_runtime_id).inspect())
+            .State.Status;
         container.docker_status = dockerStatus;
       } catch (e) {
         container.docker_status = 'not exist';
@@ -196,9 +225,10 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async changeEnvs(input: {
-    key: string,
-    value: string,
+    key: string;
+    value: string;
   }[]) {
+    await this.checkActionAccess(resources.container.actions.edit);
     await this.removeContainerEnvOptions();
     for (let i in input) {
       const env = input[i];
@@ -213,7 +243,11 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async changeVariables(input: Partial<ContainerVariableTable>[]) {
-    await knex<ContainerVariableTable>('container_variable').delete().where('container_id', this.id);
+    await this.checkActionAccess(resources.container.actions.edit);
+    await knex<ContainerVariableTable>('container_variable').delete().where(
+      'container_id',
+      this.id,
+    );
     for (let i in input) {
       await knex<ContainerVariableTable>('container_variable').insert({
         container_id: this.id,
@@ -222,55 +256,67 @@ class Container extends BaseRepository<ContainerTable> {
     }
   }
 
-  shellCommand(command: string): { id: string, output: Promise<{ stdout: string, stderr: string }> } {
+  shellCommand(
+    command: string,
+  ): { id: string; output: Promise<{ stdout: string; stderr: string }> } {
     const commandId = uuidv4();
-    return { id: commandId, output: (async () => {
-      const process = new ContainerProcess(command, this);
-      MegapolosNode.currentNode.commands[commandId] = process;
-      process.onoutput = (data) => {
-        EventsObserver.listener({ type: 'shellCommandOutput', data: data });
-      };
-      process.onerror = (data) => {
-        EventsObserver.listener({ type: 'shellCommandError', data: data });
-      };
+    return {
+      id: commandId,
+      output: (async () => {
+        const process = new ContainerProcess(command, this);
+        MegapolosNode.currentNode.commands[commandId] = process;
+        process.onoutput = (data) => {
+          EventsObserver.listener({ type: 'shellCommandOutput', data: data });
+        };
+        process.onerror = (data) => {
+          EventsObserver.listener({ type: 'shellCommandError', data: data });
+        };
 
-      await process.start();
+        await process.start();
 
-      const result = {
-        stdout: process.stdout,
-        stderr: process.stderr,
-      };
-    
-      // const result = await exec(command,
-      // // , { uid: parseInt(osUserId) }
-      // );
-      return result;
-    })() };
+        const result = {
+          stdout: process.stdout,
+          stderr: process.stderr,
+        };
+
+        // const result = await exec(command,
+        // // , { uid: parseInt(osUserId) }
+        // );
+        return result;
+      })(),
+    };
   }
 
   async getInstance(): Promise<Instance> {
     const data = await this.getData();
-    return new Instance(data.app_instance_id);
+    return new Instance(data.app_instance_id, this.userId);
   }
 
   async getImage(): Promise<Image> {
     const data = await this.getData();
-    return new Image(data.image_id);
+    return new Image(data.image_id, this.userId);
   }
 
-  async updateDockerRuntimeId(id: string):Promise<void> {
+  async updateDockerRuntimeId(id: string): Promise<void> {
     await this.edit({ docker_runtime_id: id });
   }
 
-  async getEnvs():Promise<ContainerEnvOptionTable[]> {
+  async getEnvs(): Promise<ContainerEnvOptionTable[]> {
     return this.getContainerEnvOptions();
   }
 
-  async getVariables():Promise<ContainerVariableTable[]> {
-    return knex<ContainerVariableTable>('container_variable').select('*').where('container_id', this.id);
+  async getVariables(): Promise<ContainerVariableTable[]> {
+    await this.checkActionAccess(resources.container.actions.read);
+    return knex<ContainerVariableTable>('container_variable').select('*').where(
+      'container_id',
+      this.id,
+    );
   }
 
-  async getVolumes(): Promise<{ containerVolume: ContainerVolumeTable, volume: Volume }[]> {
+  async getVolumes(): Promise<
+    { containerVolume: ContainerVolumeTable; volume: Volume }[]
+  > {
+    await this.checkActionAccess(resources.container.actions.read);
     const containerVolumes = await new Volume().getVolumesOfContainer(this.id);
     return containerVolumes.map((containerVolume) => ({
       containerVolume,
@@ -279,15 +325,20 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async getDbs(): Promise<ContainerDbTable[]> {
+    await this.checkActionAccess(resources.container.actions.read);
     return new ContainerDb().getByFields({ container_id: this.id });
   }
 
   async getDockerLog(): Promise<string> {
+    await this.checkActionAccess(resources.container.actions.read);
     const data = await this.getData();
     return new MegapolosNode(data.node_id).getDockerContainerLog(data.id);
   }
 
-  async listFiles(path: string): Promise<{ files: string[], directories: string[] }> {
+  async listFiles(
+    path: string,
+  ): Promise<{ files: string[]; directories: string[] }> {
+    await this.checkActionAccess(resources.container.actions.read);
     const output = await this.shellCommand(`ls -p ${path}`).output;
     const files: string[] = [];
     const directories: string[] = [];
@@ -296,7 +347,9 @@ class Container extends BaseRepository<ContainerTable> {
         return;
       }
       if (line.endsWith('/')) {
-        directories.push((path === '/' ? path : path + '/') + line.slice(0, -1));
+        directories.push(
+          (path === '/' ? path : path + '/') + line.slice(0, -1),
+        );
       } else {
         files.push((path === '/' ? path : path + '/') + line);
       }
@@ -305,17 +358,20 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async showFile(path: string): Promise<string> {
+    await this.checkActionAccess(resources.container.actions.read);
     const output = await this.shellCommand(`cat ${path}`).output;
     return output.stdout;
   }
 
-  async getContainerEnvOptions():Promise<ContainerEnvOptionTable[]> {
+  async getContainerEnvOptions(): Promise<ContainerEnvOptionTable[]> {
+    await this.checkActionAccess(resources.container.actions.read);
     return knex<ContainerEnvOptionTable>('container_env_option')
       .select('container_env_option.*')
       .where('container_env_option.container_id', this.id);
   }
 
   async addContainerEnvOption(input: Partial<ContainerEnvOptionTable>) {
+    await this.checkActionAccess(resources.container.actions.edit);
     await knex<ContainerEnvOptionTable>('container_env_option').insert({
       container_id: this.id,
       ...input,
@@ -323,10 +379,15 @@ class Container extends BaseRepository<ContainerTable> {
   }
 
   async removeContainerEnvOptions() {
-    await knex<ContainerEnvOptionTable>('container_env_option').delete().where('container_id', this.id);
+    await this.checkActionAccess(resources.container.actions.edit);
+    await knex<ContainerEnvOptionTable>('container_env_option').delete().where(
+      'container_id',
+      this.id,
+    );
   }
 
   async getDomain(): Promise<DomainTable | null> {
+    await this.checkActionAccess(resources.container.actions.read);
     const data = await this.getData();
     if (!data.domain_id) {
       return null;
@@ -334,7 +395,10 @@ class Container extends BaseRepository<ContainerTable> {
     return new Domain(data.domain_id).getData();
   }
 
-  async getRuntimeVariables(withoutInstance: boolean = false): Promise<ContainerRuntimeVariables> {
+  async getRuntimeVariables(
+    withoutInstance: boolean = false,
+  ): Promise<ContainerRuntimeVariables> {
+    await this.checkActionAccess(resources.container.actions.read);
     const domain = await this.getDomain();
     const volumes = await this.getVolumes();
     const variables = await this.getVariables();
@@ -379,7 +443,12 @@ class Container extends BaseRepository<ContainerTable> {
     return result;
   }
 
-  addDb(dbId: string, dbUserId: string, name: string): Promise<ContainerDbTable> {
+  async addDb(
+    dbId: string,
+    dbUserId: string,
+    name: string,
+  ): Promise<ContainerDbTable> {
+    await this.checkActionAccess(resources.container.actions.edit);
     return new ContainerDb().create({
       container_id: this.id,
       db_id: dbId,
@@ -388,10 +457,14 @@ class Container extends BaseRepository<ContainerTable> {
     });
   }
 
-  removeDb(id: string): Promise<boolean> {
+  async removeDb(id: string): Promise<boolean> {
+    await this.checkActionAccess(resources.container.actions.edit);
     return new ContainerDb(id).delete();
   }
-    
+
+  async _updateLifeStatus(status: string): Promise<void> {
+    await this.edit({ life_status: status });
+  }
 }
 
 export default Container;
