+++ classes/Image.ts
@@ -1,24 +1,32 @@
-import { v4 as uuidv4 } from 'uuid';
 import fse from 'fs-extra';
-import { ContainerTable, ImageEnvRequirementTable, ImageStatus, ImageTable, ImageVariableRequirementTable, LogTable, LogType } from '../modules/models/tables';
-import App from './App';
-import docker from '../coreDocker';
+import { v4 as uuidv4 } from 'uuid';
 import { megapolosPath } from '..';
-import Repository from './Repository';
-import MegapolosNode from './Node';
-import User from './User';
 import config from '../config/config';
-import Log from './Log';
 import { knex } from '../corePostgres';
+import {
+  ImageEnvRequirementTable,
+  ImageStatus,
+  ImageTable,
+  ImageVariableRequirementTable,
+  LogTable,
+  LogType,
+} from '../modules/models/tables';
+import { resources } from '../src/features/rights/resources_list';
+import App from './App';
 import BaseRepository from './BaseRepository';
 import Container from './Container';
+import Log from './Log';
+import MegapolosNode from './Node';
+import Repository from './Repository';
+import User from './User';
 
 class Image extends BaseRepository<ImageTable> {
   getTable(): string {
     return 'image';
   }
 
-  async build(userId: string) {
+  async build() {
+    await this.checkActionAccess(resources.image.actions.build);
     const data = await this.getData();
     if (!data.repository_id) {
       return;
@@ -27,38 +35,57 @@ class Image extends BaseRepository<ImageTable> {
     if (!await fse.exists(path)) {
       await fse.mkdir(path);
     }
-    const repository = new Repository(data.repository_id);
+    const repository = new Repository(data.repository_id, this.userId);
     await repository.fetch();
     await repository.copyBranchTo(path, data.branch);
     console.log(data);
     if (data.repository_id) {
       await this.edit({ status: ImageStatus.Building });
       try {
-        const log = new Log();
-        await log.create({ 
+        const log = new Log(undefined, this.userId);
+        await log.create({
           name: 'Build image ' + data.name,
           object_id: this.id,
           object_name: data.name,
           type: LogType.ImageBuild,
         });
 
-        let tags = `-t ${data.image} -t ${config.registryHost}:443/${data.image}`;
+        let tags =
+          `-t ${data.image} -t ${config.registryHost}:443/${data.image}`;
         if (config.devMode) {
           tags = `-t ${data.image}`;
         }
-    
+
         const result = await MegapolosNode.currentNode.shellCommand(
-          `cd ${path} && docker build ${tags} .`, new User(userId), log).output;
+          `cd ${path} && docker build ${tags} .`,
+          new User(this.userId),
+          log,
+        ).output;
         if (!config.devMode) {
-          await MegapolosNode.currentNode.shellCommand(`docker login -u '${config.registryUser}' -p '${config.registryPassword}' ${config.registryHost}:443`, new User(userId), log).output;
-          await MegapolosNode.currentNode.shellCommand(`docker push ${config.registryHost}:443/${data.image}`, new User(userId), log).output;
-          await MegapolosNode.currentNode.shellCommand(`docker image prune -f`, new User(userId), log).output;
+          await MegapolosNode.currentNode.shellCommand(
+            `docker login -u '${config.registryUser}' -p '${config.registryPassword}' ${config.registryHost}:443`,
+            new User(this.userId),
+            log,
+          ).output;
+          await MegapolosNode.currentNode.shellCommand(
+            `docker push ${config.registryHost}:443/${data.image}`,
+            new User(this.userId),
+            log,
+          ).output;
+          await MegapolosNode.currentNode.shellCommand(
+            `docker image prune -f`,
+            new User(this.userId),
+            log,
+          ).output;
         }
-        await this.edit({ status: ImageStatus.Built, last_build_date: new Date() });
+        await this.edit({
+          status: ImageStatus.Built,
+          last_build_date: new Date(),
+        });
         console.log(result);
         if (await fse.exists(path)) {
           await fse.remove(path);
-        }    
+        }
       } catch (error) {
         await this.edit({ status: ImageStatus.NotExist });
         console.error(error);
@@ -67,15 +94,16 @@ class Image extends BaseRepository<ImageTable> {
         }
       }
     }
-  }  
+  }
 
   async getApp(): Promise<App> {
     const data = await this.getData();
-    return new App(data.app_id);
+    return new App(data.app_id, this.userId);
   }
-  
+
   async updateNodes(): Promise<void> {
-    const containers = await new Container().getByFields({ image_id: this.id });
+    await this.checkActionAccess(resources.image.actions.update_nodes);
+    const containers = await new Container(undefined, this.userId).getByFields({ image_id: this.id });
     const nodes: string[] = [];
     for (let i in containers) {
       const container = containers[i];
@@ -88,6 +116,7 @@ class Image extends BaseRepository<ImageTable> {
   }
 
   async changeEnvs(envs: ImageEnvRequirementTable[]): Promise<boolean> {
+    await this.checkActionAccess(resources.image.actions.edit);
     await knex('image_env_requirement').where({ image_id: this.id }).delete();
     for (let i in envs) {
       const env = envs[i];
@@ -98,8 +127,12 @@ class Image extends BaseRepository<ImageTable> {
     return true;
   }
 
-  async changeVariables(variables: ImageVariableRequirementTable[]): Promise<boolean> {
-    await knex('image_variable_requirement').where({ image_id: this.id }).delete();
+  async changeVariables(
+    variables: ImageVariableRequirementTable[],
+  ): Promise<boolean> {
+    await this.checkActionAccess(resources.image.actions.edit);
+    await knex('image_variable_requirement').where({ image_id: this.id })
+      .delete();
     for (let i in variables) {
       const variable = variables[i];
       delete variable.id;
@@ -110,18 +143,24 @@ class Image extends BaseRepository<ImageTable> {
   }
 
   async getEnvs(): Promise<ImageEnvRequirementTable[]> {
+    await this.checkActionAccess(resources.image.actions.read);
     return knex('image_env_requirement').where({ image_id: this.id });
   }
 
   async getVariables(): Promise<ImageVariableRequirementTable[]> {
+    await this.checkActionAccess(resources.image.actions.read);
     return knex('image_variable_requirement').where({ image_id: this.id });
   }
 
   async getLastBuildLog(): Promise<LogTable> {
-    return (await new Log().getByQuery(_knex => 
-      _knex.where({ object_id: this.id, type: LogType.ImageBuild }).orderBy('create_date', 'desc').limit(1)))[0];
+    await this.checkActionAccess(resources.image.actions.read);
+    return (await new Log().getByQuery(_knex =>
+      _knex.where({ object_id: this.id, type: LogType.ImageBuild }).orderBy(
+        'create_date',
+        'desc',
+      ).limit(1)
+    ))[0];
   }
-
 }
 
 export default Image;
